集货区删除增加站点开关及businessLog日志记录

+++ bluedragon-distribution-web/src/main/java/com/jd/bluedragon/distribution/collect/controller/CollectGoodsAreaController.java
@@ -139,6 +139,11 @@ public class CollectGoodsAreaController {
 					}
 				}
 			}
+			// 没有删除权限，直接返回
+            if(collectGoodsAreaService.checkAuthority(userCode,createSiteCode,codes)){
+                rest.toFail("不可删除，可邮件express_dms@jd.com申请开启场地的删除权限!");
+                return rest;
+            }
 			rest.setData(collectGoodsAreaService.deleteByCode(codes));
 		} catch (Exception e) {
 			log.error("fail to delete！",e);
